feat: improved project creation interface. fix #22
